add client to shipment resource
